+++ NumberSequenceProvider.hpp
@@ -9,9 +9,9 @@ namespace Lights
 		inline NumberSequenceProvider() : NumberProvider( 0 ) {}
 		inline void Next() { providedValue = sequence.Next(); }
 		inline void Reset() { sequence.Reset(); }
-		inline void AddValue( uint16_t valueToAdd ) { sequence.Add( valueToAdd ); }
+		inline void AddValue( int32_t valueToAdd ) { sequence.Add( valueToAdd ); }
 
 	private:
-		ProviderSequenceHelper<uint16_t> sequence;
+		ProviderSequenceHelper<int32_t> sequence;
 	};
 }
